backend/requirements.txt
# FastAPI & server
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1

# Database
sqlalchemy==2.0.21
alembic==1.12.0

# Google API
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.97.0

# NLP & text processing
spacy==3.6.1
nltk==3.8.1
scikit-learn==1.3.0

# PDF generation
reportlab==4.0.4

# Utilities
httpx==0.24.1
aiohttp==3.8.5
tenacity==8.2.3